--- deeppavlov/configs/odqa/en_odqa_pop_infer_enwiki20180211.json
@@ -1,81 +1,0 @@
-{
-  "chainer": {
-    "in": [
-      "question_raw"
-    ],
-    "out": [
-      "best_answer"
-    ],
-    "pipe": [
-      {
-        "config_path": "{CONFIGS_PATH}/doc_retrieval/en_ranker_pop_enwiki20180211.json",
-        "in": [
-          "question_raw"
-        ],
-        "out": [
-          "pop_doc_ids"
-        ]
-      },
-      {
-        "class_name": "wiki_sqlite_vocab",
-        "in": [
-          "pop_doc_ids"
-        ],
-        "out": [
-          "pop_doc_text"
-        ],
-        "join_docs": false,
-        "shuffle": false,
-        "load_path": "{DOWNLOADS_PATH}/odqa/enwiki.db"
-      },
-      {
-        "class_name": "document_chunker",
-        "in": ["pop_doc_text"],
-        "out": ["chunks"],
-        "tokens_limit": 4000,
-        "flatten_result": true
-      },
-      {
-        "class_name": "string_multiplier",
-        "in": [
-          "question_raw",
-          "chunks"
-        ],
-        "out": [
-          "questions"
-        ]
-      },
-      {
-        "class_name": "logit_ranker",
-        "batch_size": 10,
-        "squad_model": {
-          "config_path": "{CONFIGS_PATH}/squad/qa_squad2_bert.json"
-        },
-        "sort_noans": true,
-        "in": [
-          "chunks",
-          "questions"
-        ],
-        "out": [
-          "best_answer",
-          "best_answer_score"
-        ]
-      }
-    ]
-  },
-  "train": {
-    "batch_size": 10000,
-    "evaluation_targets": [],
-    "class_name": "fit_trainer"
-  },
-  "metadata": {
-    "variables": {
-      "ROOT_PATH": "~/.deeppavlov",
-      "DOWNLOADS_PATH": "{ROOT_PATH}/downloads",
-      "MODELS_PATH": "{ROOT_PATH}/models",
-      "CONFIGS_PATH": "{DEEPPAVLOV_PATH}/configs"
-    },
-    "download": [
-    ]
-  }
-}
